+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add some additional logging when retrieving secrets.
+
 ### [v1.11](https://github.com/realityforge/keycloak-converger/tree/v1.11) (2020-09-01) · [Full Changelog](https://github.com/realityforge/keycloak-converger/compare/v1.10...v1.11)
 
 * Avoid mandating that the secrets directory exists.