+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-### Unreleased
+### [v1.12](https://github.com/realityforge/keycloak-converger/tree/v1.12) (2020-09-01) · [Full Changelog](https://github.com/realityforge/keycloak-converger/compare/v1.11...v1.12)
 
 * Add some additional logging when retrieving secrets.
 * Fix bug where client secrets were being incorrectly downloaded for public clients and not for confidential clients.